Fix back button navigation to be triggered by each fragment
