+++ app/src/main/java/com/example/barcode_a/HealthPreference.kt
@@ -6,6 +6,8 @@ import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
 import android.widget.Button
+import android.widget.Toast
+import androidx.activity.addCallback
 
 
 class HealthPreference : Fragment() {
@@ -40,6 +42,15 @@ class HealthPreference : Fragment() {
         val btnAllergies = view.findViewById<Button>(R.id.btnAllergies)
         btnAllergies.setOnClickListener {
             val fragment = Allergies()
+            parentFragmentManager.beginTransaction()
+                .replace(R.id.frame_layout, fragment)
+                .addToBackStack(null)
+                .commit()
+        }
+
+        //Back Button Function
+        requireActivity().onBackPressedDispatcher.addCallback(viewLifecycleOwner) {
+            val fragment = Home()
             parentFragmentManager.beginTransaction()
                 .replace(R.id.frame_layout, fragment)
                 .addToBackStack(null)
