+++ app/src/main/java/com/example/barcode_a/Home.kt
@@ -2,6 +2,7 @@ package com.example.barcode_a
 
 import android.content.Intent
 import android.os.Bundle
+import android.os.Handler
 import androidx.fragment.app.Fragment
 import android.view.LayoutInflater
 import android.view.View
@@ -9,6 +10,8 @@ import android.view.ViewGroup
 import android.widget.Button
 import android.widget.LinearLayout
 import android.widget.Toast
+import androidx.activity.addCallback
+import androidx.core.app.ActivityCompat.finishAffinity
 import com.budiyev.android.codescanner.CodeScannerView
 import com.example.barcode_a.databinding.FragmentHomeBinding
 import com.google.firebase.auth.FirebaseAuth
@@ -16,6 +19,9 @@ import com.google.firebase.auth.FirebaseAuth
 class Home : Fragment() {
 
     private lateinit var firebaseAuth: FirebaseAuth
+
+    private val delay : Long = 3000 // 3 seconds delay
+    var quit = false
 
 
     override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
@@ -25,6 +31,23 @@ class Home : Fragment() {
         val activity = requireActivity()
 
         firebaseAuth = FirebaseAuth.getInstance()
+
+
+        //Back Button Function
+        requireActivity().onBackPressedDispatcher.addCallback(viewLifecycleOwner) {
+            if (quit == false){
+                Toast.makeText(activity, "Press Again To Quit", Toast.LENGTH_SHORT).show()
+                quit = true
+
+                val handler = Handler()
+                handler.postDelayed({
+                    quit = false
+                }, delay)
+            }
+            else{
+                finishAffinity(activity)
+            }
+        }
 
         //Sign Out Function
         btnHomeSignOut.setOnClickListener(){
@@ -58,6 +81,8 @@ class Home : Fragment() {
                 .addToBackStack(null)
                 .commit()
         }
+
+
     }
 
     override fun onCreateView(
@@ -68,4 +93,6 @@ class Home : Fragment() {
         return inflater.inflate(R.layout.fragment_home, container, false)
 
     }
+
+
 }
