+++ app/src/main/java/com/example/barcode_a/MainActivity.kt
@@ -30,8 +30,6 @@ class MainActivity : AppCompatActivity() {
     private lateinit var firebaseAuth: FirebaseAuth
     private val CAMERA_REQUEST_CODE = 101
 
-    private val delay : Long = 3000 // 3 seconds delay
-    var quit = false
 
 
     override fun onCreate(savedInstanceState: Bundle?) {
@@ -110,19 +108,5 @@ class MainActivity : AppCompatActivity() {
     }
 
 //Quit Application
-    override fun onBackPressed() {
 
-        if (quit == false){
-            Toast.makeText(this, "Press Again To Quit", Toast.LENGTH_SHORT).show()
-            quit = true
-
-            val handler = Handler()
-            handler.postDelayed({
-                quit = false
-            }, delay)
-        }
-        else{
-            finishAffinity()
-        }
-    }
 }
