+++ app/src/main/java/com/example/barcode_a/MainActivityManufacturer.kt
@@ -75,4 +75,5 @@ class MainActivityManufacturer : AppCompatActivity() {
         }
     }
 
+
 }
